Control via headphones Adding ability to edit albumArt url Capturing volume keys Adding edit command to episode list Adding preferences (Feed details/album art/single tap options) Starting activity when downloading (hopefully keeps wifi on) Handling directory/filename for downloads Don't auto-delete files we are listening to

+++ bin/app/models/feed.js
@@ -206,7 +206,11 @@ Feed.prototype.updateCheck = function(transport, callback) {
 			if (this.autoDownload &&
 				this.maxDownloads > 0 && downloaded >= this.maxDownloads &&
 				e.downloaded) {
-				e.deleteFile();
+				if (!e.position) {
+					e.deleteFile();
+				} else {
+					downloaded++;
+				}
 			} else if (episode.downloaded || episode.downloadTicket) {
 				downloaded++;
 			}
